refactor(*): make components (#21)
This pull request includes several changes to the `src` directory to
improve code quality and functionality. The most important changes
include the addition of a new component for GitHub repository stars,
updates to the `ContributionsHandler` class to use read-only arrays, and
refactoring of the `generateContributions` function.

### New Component:
*
[`src/components/github-repo-stars.js`](diffhunk://#diff-89f47d310d097404bd1bf5467d9f1dd7fcee93567364e078481d4adc5aba30a1R1-R17):
Added a new component `GitHubRepoStars` to generate a GitHub repository
stars badge.
*
[`src/components/index.js`](diffhunk://#diff-c7db7feaa21621735b0b41422b0d1ba0aae29948113c1a3bf63575d63217b10aR1-R3):
Imported and exported the new `GitHubRepoStars` component.

### Updates to ContributionsHandler:
*
[`src/core/helpers/contributions-handler.js`](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accL27-R30):
Updated the `ContributionsHandler` class to use read-only arrays for the
`contributions` property and its getter method.
[[1]](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accL27-R30)
[[2]](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accL113-R113)

### Refactoring:
*
[`src/generators/contributions.js`](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L2-R9):
Renamed the file from `src/generator/contributions.js`, imported the
`GitHubRepoStars` component, and refactored the `generateContributions`
function to use this component and return the generated markdown instead
of writing it directly to a file.
[[1]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L2-R9)
[[2]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L28)
[[3]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L44-L53)
[[4]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L82-R80)
[[5]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L130-R128)

### Entry File:
*
[`src/index.js`](diffhunk://#diff-bfe9874d239014961b1ae4e89875a6155667db834a410aaaa2ebe3cf89820556L1-R28):
Updated the entry file to include a helper function `writeMarkdownFile`
and use it to write the generated contributions markdown to a file.