feat(*): update `contributions.md` (#24)
This pull request includes several changes to enhance the contributions
documentation and improve the functionality of the contributions
handler. The most important changes include updating the contributions
documentation format, adding new methods to handle various types of
contributions, and improving the type-to-title transformation.

Improvements to contributions documentation:

*
[`docs/contributions.md`](diffhunk://#diff-88440b1d44a319d356349d43ee7f53428faa1c7fd50d75b799719c7c231857b3L14-L26):
Updated the format to include new sections for pull requests, issues,
and comments.
*
[`src/generators/contributions.js`](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L47-R47):
Modified the generation of `contributions.md` to reflect the new format
and added counts for different types of contributions.
[[1]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L47-R47)
[[2]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153L62-L73)
[[3]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153R94)
[[4]](diffhunk://#diff-7f660510ea8b99897096ec7dcd24f2c5530b48f2c52cb1104f43df6563f10153R108)

Enhancements to contributions handler:

*
[`src/core/helpers/contributions-handler.js`](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accR13-R17):
Added new methods to handle issues, pull request comments, issue
comments, and discussion comments.
[[1]](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accR13-R17)
[[2]](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accL49-R150)
[[3]](diffhunk://#diff-82d24185ce8003662de9eb7b74fc46081f2b735c20213cba4f3991bf93662accL88-R214)

Improvements to type-to-title transformation:

*
[`src/core/helpers/type-to-title.js`](diffhunk://#diff-ff315a9524ea6cf6798db006aa1fb8d6c2fdf3864652302495d2a745434ae0b7R13-R72):
Added an `Options` parameter to customize the output of the
`typeToTitle` function, allowing for optional inclusion of emojis and
titles.

Minor fixes:

*
[`src/core/helpers/sort-contributions.js`](diffhunk://#diff-8c4b92adaa39a056679372847bfb18b45aa286f43c5ecf759ad372e4482b114fL32):
Removed the 'revert' type from the list of recognized contribution
types.
*
[`src/data/contributions.js`](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L743):
Removed TODO comments and ensured empty pull request arrays are not
counted as contributions.
[[1]](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L743)
[[2]](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L755)
[[3]](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L1036)