feat(*): add merged PRs (#32)
This pull request includes updates to the documentation and
contributions data, reflecting recent changes in the project. The most
important changes include updates to the contributions summary in
`docs/contributions.md` and the status of contributions in
`src/data/contributions.js`.

Updates to documentation:

*
[`docs/contributions.md`](diffhunk://#diff-88440b1d44a319d356349d43ee7f53428faa1c7fd50d75b799719c7c231857b3L18-R22):
Updated the number of merged pull requests and various contribution
statistics.
*
[`docs/contributions.md`](diffhunk://#diff-88440b1d44a319d356349d43ee7f53428faa1c7fd50d75b799719c7c231857b3L205-R214):
Corrected the categorization of a feature and a chore.

Updates to contributions data:

*
[`src/data/contributions.js`](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L703-R703):
Marked the `applyInlineConfig` feature as merged.
*
[`src/data/contributions.js`](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L728-R730):
Changed the type and status of the `PULL_REQUEST_TEMPLATE.md`
contribution to a chore and marked it as merged.