feat(*): update contributions (#52)
This pull request includes minor updates to contribution data and
documentation to reflect recent changes in the project. The updates
include adjustments to contribution counts, a status update for a merged
pull request, and a correction in the emoji used for a specific entry.

### Updates to contribution data:

*
[`src/data/contributions.js`](diffhunk://#diff-7cdf3ef9fb7dd80c0bda04d785a541f4e6e0b2c0786ddf469561646eef9ebec1L762-R762):
Updated the `merged` status of pull request #345 from `false` to `true`.

### Documentation updates:

* `docs/contributions.md`: 
- Updated the total count of merged pull requests from "More than 108"
to "More than 109" and incremented the count of `chore` contributions
from 17 to 18.
- Corrected the emoji for pull request #345 from :green_heart: to
:purple_heart:.