chore(*): move `FUNDING.yml` to `.github` repository (#105)
This pull request removes references to the `FUNDING.yml` file and its
associated configurations. The changes simplify the repository by
eliminating unused or unnecessary funding-related configurations.

### Removal of funding configurations:

*
[`.github/FUNDING.yml`](diffhunk://#diff-07985fdcade0e64d11482724879a644f07879ba61b8fb6c6119e1b1902b72ae4L1-L2):
Deleted the GitHub and Open Collective funding links for `lumirlumir`.
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L14-L15):
Removed the synchronization configuration for the `FUNDING.yml` file,
ensuring it is no longer copied to the `configs/.github` directory.